Move shared python code to a shared directory
I have implemented these changes for the dome test case; the same
changes are needed in the other test cases, as well as tests/new/
